+++ CHANGELOG.md
@@ -179,5 +179,6 @@
 - Studio : Fix error when click on wkf buttons.
 - Studio : Fix export app.
 - Advance Data Import : Fix indexOutOfBound Exception.
+- Advanced Data Import : Specify truncated value to sampleLines for large string.
 
 [Unreleased 5.2.0]: https://github.com/axelor/axelor-business-suite/compare/dev...wip